Make SM2 functions private
Address issue #5670
