+++ crypto/sm2/sm2_err.c
@@ -9,7 +9,7 @@
  */
 
 #include <openssl/err.h>
-#include <openssl/sm2err.h>
+#include "internal/sm2err.h"
 
 #ifndef OPENSSL_NO_ERR
